requirements.txt
# Core
streamlit>=1.29.0
pandas>=2.1.0
numpy>=1.24.0

# Visualização
plotly>=5.18.0

# Google Sheets
gspread>=5.12.0
google-auth>=2.25.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1

# Utilidades
requests>=2.31.0
pytz>=2023.3
python-dateutil>=2.8.2

# Cache e Performance
streamlit-extras>=0.3.5

# Segurança
python-dotenv>=1.0.0

# Desenvolvimento (opcional)
black>=23.0.0
flake8>=6.0.0
pytest>=7.4.0